chore(android): Refactor HelpFile as utility
